Move the wayland backend into the app_tray module
This PR moves the wayland backend and the app tray config into the AppTray struct in the app_tray module. This is much cleaner than having it in the overarching Panel struct. Also looking into turning the AppTray struct into an iced component, as it's relatively self-contained, and should only need a config passed into it

Signed-off-by: Ryan Brue <ryanbrue.dev@gmail.com>